Implement GetDrivesAsync. ab#51

+++ Functionland.FxFiles/Shared/Models/FsArtifact.cs
@@ -16,7 +16,7 @@ namespace Functionland.FxFiles.Shared.Models
         public FsFileProviderType? ProviderType { get; set; }
         public string? MimeType { get; set; }
         public long? Size { get; set; }
-        public int? Capacity { get; set; }
+        public long? Capacity { get; set; }
         public string? ContentHash { get; set; }
         public int? ParentId { get; set; }
         public string? OriginDevice { get; set; }
